Run query once per search It looks like controllers/catalog/Search.php is running each search query twice: once for the items on the current page, and then again to count the total number of matches for the search.
There doesn't seem to be a clean way to count the results, but here's
_one_ way.  We bundle up the whole bunch of `UNION` sets as a sub-query,
and we do a count(*) on that set.  We can't do just a _regular_ count(*),
or it would return our limit of 25.

I'm not sure just how much of a hack this `OVER()` call[1][2] is, but it
at least seems faster than running the entire query twice.

[1]: https://stackoverflow.com/a/8242764
[2]: https://stackoverflow.com/a/28888696