Tests for advanced_search
I wanted to set a precedent for how to write search tests, so here are
two tests that:

* Test that we can get an empty result
* Test that we can do a very simple title search

These aren't really "unit" tests in that they do a full test of
everything from the controllers, to the models, to the views. I might
try writing some more granular tests in the future, but, for now,
something is better than nothing.

Another problem with these tests is that we don't reset the database
between tests or anything, so it could pollute peoples local databases.
(It doesn't bother me, I reset my db all the time, but others might
not.)