Add manual sort functionality for genres
New feature:
Genre Manager now shows an editable 'Sort Order' box for each genre.
This is not a global number - it is compared only to 'sibling' genres
 within the same (sub-)menu.  Lower numbers are sorted first, with
 alphabetical sorting as the tie-breaker.

This ordering is honored in the volunteer-facing application:
* Genre Manager (both main editor and its 'Parent' drop-downs)
* Tree drop-down in Template Generator
* Tree drop-down in admin Project Screen (add_catalog_item)
Sort order in the public catalog is left unchanged for now.

We also fix a bug where a genre's `parent_id` could be set to its _own_
 ID, or the ID of one of its sub-items.

A minor bug not fixed: `mahana_hierarchy->resync()` tries to update the
 `deep` property of sub-items, but only does one pass.  Depending on ID
 ordering, it can take several clicks of the 'Save' button before nested
 items are all up to date.